backend/requirements.txt
# backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
python-jose[cryptography]
bcrypt==3.2.0
passlib
python-multipart
psycopg2-binary
python-dotenv